services/authenticator/requirements.txt
pyjwt[crypto]
argon2-cffi
fastapi
httpx
psycopg[binary]